requirements.txt
telethon==1.28.5
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pandas==2.1.3
matplotlib==3.8.2
transformers==4.35.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.0+cpu
pyshorteners==1.0.1 